requirements.txt
imbalanced-learn==0.14.1
lazy-loader==0.5
lightgbm==4.6.0
lime==0.2.0.1
matplotlib==3.10.8
matplotlib-inline==0.2.1
numpy==2.4.3
pandas==2.3.3
pytest==9.0.3
pytest-asyncio==1.3.0
pytest-cov==7.1.0
reportlab==4.4.10
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
shap==0.51.0
sklearn-compat==0.1.5
tifffile==2026.4.11
ucimlrepo==0.0.7
xgboost==3.2.0
